done fixing image not show properly
