+++ src/components/data/menu.tsx
@@ -26,7 +26,7 @@ export const linksMenu = [
     icon: (
       <IconNotes className="h-full w-full text-neutral-500 dark:text-neutral-300" />
     ),
-    href: "/note",
+    href: "/notes",
   },
   {
     title: "Projects",
